Remove prosemirror-schema-basic dependency
I was trying to setup an installation of prosemirror without using the
`prosemirror-schema-basic` module and `npm install` errored out due to
not being able to access the `prosemirror-schema-basic` module. I
couldn't find a place that `prosemirror-example-setup` actually uses
the `prosemirror-schema-basic`, so this commit removes the unused
dependency from the `package.json` file.